// src/api/types/diff_line.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DiffLineKind } from "./diff_line_kind";

export type DiffLine = { kind: DiffLineKind, old_no: number | null, new_no: number | null, text: string, no_eol: boolean, has_crlf: boolean, };